% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_using}
\alias{read_using}
\title{Read from S3 using a particular function}
\usage{
read_using(FUN, s3_path, ...)
}
\arguments{
\item{FUN}{a function to parse the data into}

\item{s3_path}{path to the s3 file bucket/folder/file.txt}

\item{...}{arguments for said function}
}
\value{
whatever the function returns
}
\description{
Read from S3 using a particular function
}
\examples{
Rs3tools::read_using(FUN=readxl::read_excel, s3_path="alpha-test-team/mpg.xlsx")
}
