% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{s3_path_to_full_df}
\alias{s3_path_to_full_df}
\title{Read a full file from S3, using the full path to the file including the
bucketname.}
\usage{
s3_path_to_full_df(s3_path, ...)
}
\arguments{
\item{path}{a string -  the full path to the file including the bucketname}
}
\value{
a dataframe (.csv, .tsv), tibble (Excel, SAS, SPSS, Stata), or file location
}
\description{
This function will attempt to read the file directly, as a dataframe.
If this is not possible it will download the file to a temporary location
and load it.
At present the function supports direct reading of CSV, TSV, XLS, XLSX,
SAS7DBAT, SAV and DTA file types.
You can add options to the read function that are compatible with
readxl::read_excel() and read.csv(). See their help files for more info.
}
\examples{
df <- s3_read_path_to_df("alpha-moj-analytics-scratch/folder/file.csv")
df <- s3_read_path_to_df("alpha-moj-analytics-scratch/folder/file.tsv")
df <- s3_read_path_to_df("alpha-moj-analytics-scratch/folder/file.xls")
df <- s3_read_path_to_df("alpha-moj-analytics-scratch/folder/file.xls", sheet = 1)
filelocation <- s3_read_path_to_df("alpha-moj-analytics-scratch/folder/file.png")
}
