% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{s3_svc}
\alias{s3_svc}
\title{Obtain a paws S3 client using MoJ AP defaults.}
\usage{
s3_svc(region = REGION, ...)
}
\value{
paws::s3 client
}
\description{
This function should be called each time you would use a paws::s3 client. It
sets the default region and handles authentication for the Analytical
Platform.
}
\examples{
Rs3tools::s3_svc()$put_object(Body = "filetoupload", Bucket = "examplebucket", Key = "objectkey")
}
